Move most of the logic into game-logic subcrate
