+++ Cargo.toml
@@ -22,6 +22,7 @@ rand = "0.8.5"
 getrandom = { version = "0.2.12", features = ["js"] }
 serde = { version = "1.0", features = ["derive"] }
 serde-wasm-bindgen = "0.6.3"
+asteroid-colonies-logic = { path = "game-logic" }
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3.34"
