+++ game-logic/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "asteroid-colonies-logic"
+version = "0.1.0"
+edition = "2021"
+description = "Common logic between the server and the frontend"
+
+[dependencies]
+cgmath = { version = "0.18.0", features = ["serde"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+rand = "0.8.5"
+smallvec = "1.8"
+anyhow = "1.0.56"
