+++ game-logic/src/crew.rs
@@ -6,7 +6,7 @@ use crate::{
     construction::Construction,
     task::{GlobalTask, EXCAVATE_ORE_AMOUNT, LABOR_EXCAVATE_TIME},
     transport::find_path,
-    AsteroidColonies, Cell, CellState, ItemType, Pos, WIDTH,
+    AsteroidColoniesGame, Cell, CellState, ItemType, Pos, WIDTH,
 };
 
 #[derive(Clone, Debug)]
@@ -20,7 +20,7 @@ enum CrewTask {
 }
 
 #[derive(Debug)]
-pub(crate) struct Crew {
+pub struct Crew {
     pub pos: Pos,
     pub path: Option<Vec<Pos>>,
     pub from: Pos,
@@ -236,7 +236,7 @@ impl Crew {
     }
 }
 
-impl AsteroidColonies {
+impl AsteroidColoniesGame {
     pub(super) fn process_crews(&mut self) {
         let try_return = |crew: &mut Crew, buildings: &mut [Building]| {
             if let Some(building) = buildings.iter_mut().find(|b| b.pos == crew.from) {
